GitMongo 1.2

+++ src/transports/winhttp.c
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) the libgit2 contributors. All rights reserved.
+ * Copyright (C) 2009-2012 the libgit2 contributors
  *
  * This file is part of libgit2, distributed under the GNU GPL v2 with
  * a Linking Exception. For full terms see the included COPYING file.
@@ -298,7 +298,7 @@ static int write_chunk(HINTERNET request, const char *buffer, size_t len)
 		return -1;
 
 	if (!WinHttpWriteData(request,
-		git_buf_cstr(&buf),	(DWORD)git_buf_len(&buf),
+		git_buf_cstr(&buf),	git_buf_len(&buf),
 		&bytes_written)) {
 		git_buf_free(&buf);
 		giterr_set(GITERR_OS, "Failed to write chunk header");
@@ -309,7 +309,7 @@ static int write_chunk(HINTERNET request, const char *buffer, size_t len)
 
 	/* Chunk body */
 	if (!WinHttpWriteData(request,
-		buffer, (DWORD)len,
+		buffer, len,
 		&bytes_written)) {
 		giterr_set(GITERR_OS, "Failed to write chunk");
 		return -1;
@@ -494,7 +494,7 @@ replay:
 
 	if (!WinHttpReadData(s->request,
 		(LPVOID)buffer,
-		(DWORD)buf_size,
+		buf_size,
 		&dw_bytes_read))
 	{
 		giterr_set(GITERR_OS, "Failed to read data");
@@ -580,7 +580,7 @@ static int put_uuid_string(LPWSTR buffer, DWORD buffer_len_cch)
 
 static int get_temp_file(LPWSTR buffer, DWORD buffer_len_cch)
 {
-	size_t len;
+	int len;
 
 	if (!GetTempPathW(buffer_len_cch, buffer)) {
 		giterr_set(GITERR_OS, "Failed to get temp path");
@@ -639,7 +639,7 @@ static int winhttp_stream_write_buffered(
 		}
 	}
 
-	if (!WriteFile(s->post_body, buffer, (DWORD)len, &bytes_written, NULL)) {
+	if (!WriteFile(s->post_body, buffer, len, &bytes_written, NULL)) {
 		giterr_set(GITERR_OS, "Failed to write to temporary file");
 		return -1;
 	}
@@ -697,7 +697,7 @@ static int winhttp_stream_write_chunked(
 	}
 	else {
 		/* Append as much to the buffer as we can */
-		int count = min(CACHED_POST_BODY_BUF_SIZE - s->chunk_buffer_len, (int)len);
+		int count = min(CACHED_POST_BODY_BUF_SIZE - s->chunk_buffer_len, len);
 
 		if (!s->chunk_buffer)
 			s->chunk_buffer = git__malloc(CACHED_POST_BODY_BUF_SIZE);
@@ -717,7 +717,7 @@ static int winhttp_stream_write_chunked(
 			/* Is there any remaining data from the source? */
 			if (len > 0) {
 				memcpy(s->chunk_buffer, buffer, len);
-				s->chunk_buffer_len = (unsigned int)len;
+				s->chunk_buffer_len = len;
 			}
 		}
 	}
